requirements.txt
django
markdown
psycopg2
python-dotenv
setuptools
djangorestframework
djangorestframework-simplejwt
django-cors-headers
